Move device perf helpers to optimization package
