+++ pkg/azureconstants/azure_constants.go
@@ -45,22 +45,26 @@ const (
 	MinimumDiskSizeGiB            = 1
 	NetworkAccessPolicyField      = "networkaccesspolicy"
 	NotFound                      = "NotFound"
+	PerfProfileBasic              = "basic"
 	PerfProfileField              = "perfprofile"
+	PerfProfileNone               = "none"
+	PremiumAccountPrefix          = "premium"
 	PvcNameKey                    = "csi.storage.k8s.io/pvc/name"
 	PvcNamespaceKey               = "csi.storage.k8s.io/pvc/namespace"
 	PvcNamespaceTag               = "kubernetes.io-created-for-pvc-namespace"
 	PvcNameTag                    = "kubernetes.io-created-for-pvc-name"
-	PvNameTag                     = "kubernetes.io-created-for-pv-name"
 	PvNameKey                     = "csi.storage.k8s.io/pv/name"
+	PvNameTag                     = "kubernetes.io-created-for-pv-name"
 	RateLimited                   = "rate limited"
+	RequestedSizeGib              = "requestedsizegib"
 	ResizeRequired                = "resizeRequired"
 	ResourceGroupField            = "resourcegroup"
 	ResourceNotFound              = "ResourceNotFound"
-	RequestedSizeGib              = "requestedsizegib"
 	SkuNameField                  = "skuname"
 	SourceDiskSearchMaxDepth      = 10
 	SourceSnapshot                = "snapshot"
 	SourceVolume                  = "volume"
+	StandardSsdAccountPrefix      = "standardssd"
 	StorageAccountTypeField       = "storageaccounttype"
 	TagsField                     = "tags"
 	ThrottlingKey                 = "throttlingKey"
